+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/static-build
 
+## 2.9.24
+
+### Patch Changes
+
+- @vercel/gatsby-plugin-vercel-builder@2.2.2
+
+## 2.9.23
+
+### Patch Changes
+
+- @vercel/gatsby-plugin-vercel-builder@2.2.1
+
 ## 2.9.22
 
 ### Patch Changes